requirements.txt
pytest
python-dotenv
black